# chore: replace blake2b implementation by golang.org/x/crypto