--- a/version.json
+++ b/version.json
@@ -1,3 +1,3 @@
 {
-	"version": "v0.1.0"
+	"version": "v0.2.0"
 }