@marten-seemann 2022-06-02
> not blocking here, but it'd be nice if there was some verification that the hashes from the different blake2 libraries are the same

That would need to live in a separate Go module though, otherwise the dependency would stay in go.mod. Or maybe a one-time test would be enough to convince us that the libraries are doing the same?