@Jorropo 2022-06-02
I'm surprised there is no blake2 tests, Ill ad some, likely blake2's official test set.