@Jorropo 2022-06-02
I dont think we actually need to check boths, blake2 is a spec if an implementation isnt following it its wrong and that likely a big security issue.

The tests can compare to well known values.