@Jorropo 2022-06-03
/cc @aschmahmann I have added blake2b tests, tested both on the old and new implementation they both work.

RFM on my side