@Jorropo 2022-06-03
> I trust golang/x/crypto to implement Blake2 correctly.

Fair enough, I belive the goal was more to ensure the new registering is done correctly.
Maybe one used to take bits now it takes byte sizes, or something weird like that.