@Jorropo 2022-06-22
@Kubuxu 
Sorry for the tests I had fixed it locally it but forgot to push it.

I don't know if it's waiting on you, you are just the only one with merge permissions in the conversation afait.